+++ src/baldr/accessrestriction.cc
@@ -4,10 +4,10 @@
 #include <string>
 
 namespace {
-constexpr std::array<const char*, 10> kTypeToString = {
-    "hazmat",        "max_height",    "max_width",    "max_length",          "max_weight",
-    "max_axle_load", "timed_allowed", "timed_denied", "destination_allowed", "max_axles",
-};
+constexpr std::array<const char*, 11> kTypeToString =
+    {"hazmat",        "max_height",    "max_width",    "max_length",          "max_weight",
+     "max_axle_load", "timed_allowed", "timed_denied", "destination_allowed", "max_axles",
+     "vignette"};
 }
 
 namespace valhalla {
@@ -61,9 +61,10 @@ void AccessRestriction::set_value(const uint64_t v) {
   value_ = v;
 }
 
+// Serialize to JSON
 void AccessRestriction::json(rapidjson::writer_wrapper_t& writer) const {
   std::string restriction_type = "unsupported";
-  if (static_cast<size_t>(type()) < 10) {
+  if (static_cast<size_t>(type()) < kTypeToString.size()) {
     restriction_type = kTypeToString[static_cast<size_t>(type())];
   }
 
@@ -96,6 +97,9 @@ void AccessRestriction::json(rapidjson::writer_wrapper_t& writer) const {
       break;
     case AccessType::kHazmat:
       writer("value", static_cast<bool>(value()));
+      break;
+    case AccessType::kVignette:
+      writer("value", value());
       break;
     default:
       writer.set_precision(2);