+++ src/mjolnir/pbfgraphparser.cc
@@ -1164,6 +1164,13 @@ struct graph_parser {
       osmdata_.access_restrictions.insert(
           AccessRestrictionsMultiMap::value_type(osmid_, restriction));
     };
+    tag_handlers_["vignette"] = [this]() {
+      OSMAccessRestriction restriction;
+      restriction.set_type(AccessType::kVignette);
+      restriction.set_modes(kVehicularAccess);
+      osmdata_.access_restrictions.insert(
+          AccessRestrictionsMultiMap::value_type(osmid_, restriction));
+    };
     tag_handlers_["hov_type"] = [this]() {
       // If this tag is set then the way is either HOV-2 or HOV-3.
       // There are no other real-world hov levels.